refactor: Merge datahub_client into the main project
We no longer need to build this as a separate package, so we can lose
the separate pyproject.toml and combine the test suites.
