+++ core/middleware.py
@@ -1,10 +1,11 @@
 import logging
 
-from data_platform_catalogue.client.exceptions import ConnectivityError
 from django.conf import settings
 from django.core.exceptions import BadRequest
 from django.http import Http404
 from django.shortcuts import render
+
+from datahub_client.client.exceptions import ConnectivityError
 
 logger = logging.getLogger(__name__)
