+++ datahub_client/client/graphql/getChartDetails.graphql
