+++ datahub_client/client/graphql/getContainerDetails.graphql
