+++ datahub_client/client/graphql/getDashboardDetails.graphql
