+++ datahub_client/client/graphql/getDatasetDetails.graphql
