+++ datahub_client/client/graphql/getTags.graphql
