+++ datahub_client/client/search/filters.py
@@ -1,6 +1,6 @@
 from typing import Sequence
 
-from data_platform_catalogue.search_types import MultiSelectFilter
+from datahub_client.search_types import MultiSelectFilter
 
 
 def map_filters(filters: Sequence[MultiSelectFilter] | None, entity_filters=[]):
