+++ datahub_client/entities.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from enum import Enum
 from typing import Literal, Optional
 
-from pydantic import BaseModel, Field, EmailStr
+from pydantic import BaseModel, EmailStr, Field
 
 DATAHUB_DATE_FORMAT = "%Y%m%d"
