+++ home/service/base.py
@@ -1,5 +1,6 @@
-from data_platform_catalogue.client.datahub_client import DataHubCatalogueClient
 from django.conf import settings
+
+from datahub_client.client.datahub_client import DataHubCatalogueClient
 
 
 class GenericService:
@@ -8,4 +9,3 @@ class GenericService:
         return DataHubCatalogueClient(
             jwt_token=settings.CATALOGUE_TOKEN, api_url=settings.CATALOGUE_URL
         )
-
