+++ home/service/subject_area_fetcher.py
@@ -1,5 +1,6 @@
-from data_platform_catalogue.search_types import SubjectAreaOption
 from django.core.cache import cache
+
+from datahub_client.search_types import SubjectAreaOption
 
 from .base import GenericService
