--- lib/datahub-client/.gitignore
@@ -1,10 +1,0 @@
-.env
-coverage/
-venv/
-env/
-.DS_STORE
-.vscode
-*.code-workspace
-dist/
-__pycache__
-.idea/
