--- lib/datahub-client/pyproject.toml
@@ -1,29 +1,0 @@
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.isort]
-profile = "black"
-
-[tool.pylint]
-max-line-length = 120
-
-[tool.poetry]
-name = "ministryofjustice-data-platform-catalogue"
-version = "1.1.0"
-description = "Wrapper around Datahub supporting custom properties for the Find MoJ Data service."
-authors = ["MoJ Data Platform Team <data-platform-tech@digital.justice.gov.uk>"]
-license = "MIT"
-readme = "README.md"
-packages = [{ include = "data_platform_catalogue" }]
-
-[tool.poetry.dependencies]
-python = "^3.10"
-acryl-datahub = { extras = ["datahub-rest"], version = "^0.12.1.3" }
-freezegun = "^1.4.0"
-deepdiff = "^6.7.1"
-pydantic = {extras = ["email"], version = "^2.10.5"}
-
-[tool.poetry.group.dev.dependencies]
-requests-mock = "^1.11.0"
-pytest = "^7.4.2"
