+++ pyproject.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 description = ""
 authors = ["Mitch Dawson <Mitchell.Dawson@justice.gov.uk>"]
 readme = "README.md"
+package-mode = false
 
 [tool.poetry.dependencies]
 python = "^3.11"
@@ -17,16 +18,17 @@ python-dotenv = "^1.0.1"
 markdown-headdown = "^0.1.3"
 nltk = "^3.9.1"
 sentry-sdk = { extras = ["django"], version = "^2.19.2" }
-ministryofjustice-data-platform-catalogue = { path = "lib/datahub-client", develop = true }
 django-azure-auth = "2.2.0"
 django-waffle = "^4.2.0"
 psycopg = "^3.2.3"
 psycopg-binary = "^3.2.3"
 django-debug-toolbar = "^4.4.6"
-redis = {extras = ["hiredis"], version = "^5.2.1"}
+redis = { extras = ["hiredis"], version = "^5.2.1" }
 notifications-python-client = "^10.0.1"
-pydantic = {extras = ["email"], version = "^2.10.5"}
+pydantic = { extras = ["email"], version = "^2.10.5" }
 email-validator = "2.2.0"
+acryl-datahub = { extras = ["datahub-rest"], version = "^0.12.1.3" }
+deepdiff = "^6.7.1"
 
 [tool.poetry.group.dev.dependencies]
 black = "^24.10.0"
@@ -37,6 +39,7 @@ pytest-django = "^4.9.0"
 pytest-cov = "^6.0.0"
 faker = "^33.3.1"
 isort = "^5.13.2"
+freezegun = "^1.4.0"
 
 [build-system]
 requires = ["poetry-core"]
@@ -49,7 +52,7 @@ markers = [
     "slow: marks tests as slow (deselect with '-m \"not slow\"')",
     "datahub: marks tests that query Datahub rather than mocking",
 ]
-testpaths = ["tests", "lib/data_platform_catalogue/tests"]
+testpaths = ["tests"]
 
 [tool.isort]
 profile = "black"
