+++ tests/core/test_middleware.py
@@ -1,10 +1,10 @@
 from unittest.mock import MagicMock
 
-from data_platform_catalogue.client.exceptions import ConnectivityError
 from django.core.exceptions import BadRequest
 from django.http import Http404
 
 from core.middleware import CustomErrorMiddleware
+from datahub_client.client.exceptions import ConnectivityError
 
 
 def test_middleware_renders_connectivity_response():
