+++ tests/datahub_client/snapshots/test_upsert_table.json
